pyproject.toml
[project]
name = "memo-backend"
version = "0.1.0"
description = "fastapi backend for handling memos"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "pyclean>=3.1.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.3",
]
